feat: move streak to stats box
